+++ aneel.egg-info/PKG-INFO
@@ -1,0 +1,23 @@
+Metadata-Version: 1.0
+Name: aneel
+Version: 0.0.4
+Summary: This package performs various calculations related to ANEEL regulations.
+Home-page: https://github.com/renatoefsousa/ANEEL
+Author: Renato Eduardo Farias de Sousa
+Author-email: renato.ef.sousa@gmail.com
+License: MIT License
+Description: (Very short) Tutorial
+        =====================
+        First create a variable::
+            my_consumer = CompensationContinuityConsumer()
+        Adjust the class parameters::
+            my_consumer.set_features("abc", 2016, "BT", "URB", "INT", 1111, [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12],
+                         [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12], [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12],
+                         [40, 42, 43, 44, 45, 56, 47, 38, 49, 40, 51, 62])
+        Then write the result::
+            print(my_consumer)
+        Reference documentation
+        =======================
+        See https://github.com/renatoefsousa/ANEEL
+Keywords: aneel cálculos regulamentos
+Platform: UNKNOWN
