+++ aneel.egg-info/requires.txt
@@ -1,0 +1,2 @@
+csv
+json
