+++ aneel/compensation/dados2016-08-14-19-49-22-278352.csv
@@ -1,0 +1,2 @@
+"abc";2016;"BT";"URB";"INT";1111;1.0;2.0;3.0;4.0;5.0;6.0;7.0;8.0;9.0;10.0;11.0;12.0;1.0;2.0;3.0;4.0;5.0;6.0;7.0;8.0;9.0;10.0;11.0;12.0;1.0;2.0;3.0;4.0;5.0;6.0;7.0;8.0;9.0;10.0;11.0;12.0;40.45;42.34;43.23;44.76;45.12;56.23;47.78;38.65;49.98;40.56;51.34;62.65
+"def";2016;"MT";"URB";"INT";2222;1.0;2.0;3.0;4.0;5.0;6.0;7.0;8.0;9.0;10.0;11.0;12.0;1.0;2.0;3.0;4.0;5.0;6.0;7.0;8.0;9.0;10.0;11.0;12.0;1.0;2.0;3.0;4.0;5.0;6.0;7.0;8.0;9.0;10.0;11.0;12.0;40123.45;42123.87;43123.65;44123.98;45123.12;56123.43;47123.56;38123.45;49123.34;40123.76;51123.98;62123.01
